make fabs explicitly float

+++ panda/src/testbed/pgrid.cxx
@@ -249,7 +249,7 @@ load_gridded_models(WindowFramework *window,
   }
   
   grid_pos_offset = -gridwidth*GRIDCELLSIZE/2.0;
-  wander_area_pos_offset = -max(fabs(grid_pos_offset), MIN_WANDERAREA_DIMENSION/2.0f);
+  wander_area_pos_offset = -max((float)fabs(grid_pos_offset), MIN_WANDERAREA_DIMENSION/2.0f);
 
   // Now walk through the list again, copying models into the scene
   // graph as we go.
